+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## [Unreleased]
 
+### Fixed
+
+- Fix native extension loader for precompiled gems (#61)
+
 ## [0.5.1] - 2026-01-01
 
 ### Fixed